requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1
sqlalchemy==2.0.25
alembic==1.13.1
python-dateutil==2.8.2
pydantic==2.5.3
python-telegram-bot==20.3
apscheduler==3.10.1